+++ pyproject.toml
@@ -7,13 +7,13 @@ authors = [
 ]
 license = {text = "MIT"}
 readme = "README.md"
-requires-python = ">=3.12"
+requires-python = ">=3.12,<4.0"
 dependencies = [
     "pandera (==0.23.0)",
-    "gower (>=0.1.2,<0.2.0)"
+    "gower (>=0.1.2,<0.2.0)",
+    "rgs_interface @ git+https://github.com/dabadav/rgs-interface.git@v0.2.1"
 ]
 packages = [{ include = "ai_cdss", from = "src" }]
-
 
 [build-system]
 requires = ["poetry-core>=2.0.0,<3.0.0"]
