--- src/ai_cdss/recommend.py
@@ -1,19 +1,0 @@
-from ai_cdss.services.pipeline import CDSS
-
-
-def main():
-    
-    PATIENT_LIST = [
-        775,  787,  788, 1123, 1169, 1170, 1171, 1172, 1173, 1983, 2110, 2195,
-        2955, 2956, 2957, 2958, 2959, 2960, 2961, 2962, 2963, 3081, 3229, 3318, 3432
-    ]
-
-    pipeline = CDSS(
-        PATIENT_LIST,
-        n=5
-    )
-    
-    pipeline.recommend()
-
-if __name__ == "__main__":
-    main()
